Search fix + refactor (cases and units) (#678)
* wip: case search bar

* feat: search bar in case-explorer

* chore: clean up code

* testing

* fix: remove console.log

* fix: desktop version case search enabled

* feat: redux for sibling component communication

* chores: clean up, remove unnecessary function, travis error

* fix: receive correct param in memoize function

* chore: remove unnecessary data passing

* fix: only pass search term via redux
feat: display all results when search term removed
chore: removed commented out code

* fix: replace regex with indexOf

* fix: prefer all cases to search subset in non-search mode

* fix: display no unit instead of all units when no matching units

* feat: refactored Sophie's search code

The new code has clearer separation of concerns, and fixes functionality issues that existed in the previous code submission. This is applied to both the cases' and units' search functions. The new search detects whether the user is searching for a word or a number and switches search strategy accordingly.